Improve home page (#6)
* Update pubspec.yaml to include url_launcher dependency, modify pubspec.lock for version updates, and register url_launcher plugins for Linux, macOS, and Windows platforms.

* Enhance HomePage with bottom sheet functionality and URL attribution. Added a new button to expand the bottom sheet when collapsed, improved state management for bottom sheet visibility, and integrated OpenStreetMap attribution with a clickable link.

* Add places data fetching to HomePage

Introduced state management for places data, including loading and error handling. Replaced the previous FutureBuilder with a more efficient method to fetch and display places on initialization, enhancing user experience during data retrieval.

* Update dependencies and enhance HomePage with location marker functionality

Added flutter_map_location_marker dependency to pubspec.yaml and updated pubspec.lock. Enhanced HomePage to utilize CurrentLocationLayer for displaying the user's location on the map, improving the user experience with real-time location updates and a new button for aligning the map position.

* Refactor HomePage to improve bottom sheet and FAB functionality

Replaced boolean state variables for bottom sheet with a single double value to manage its extent. Introduced a ValueNotifier for the Floating Action Button (FAB) position, ensuring it updates dynamically based on the bottom sheet's state. Enhanced the layout to position the FAB correctly and removed deprecated methods for better code clarity.

* Ensure proper state management in _fetchPlacesData method by checking if the widget is mounted before updating state. This prevents potential errors during asynchronous operations.

* Refactor bottom sheet and FAB positioning in HomePage

Updated the logic for determining the bottom sheet's collapsed state by introducing a constant for minimum child size. Adjusted the Floating Action Button (FAB) positioning to a fixed value for improved layout consistency. Enhanced the DraggableScrollableSheet's minimum child size to align with the new constant, ensuring better user experience during interactions.

* Update alignPositionStreamController value in HomePage to improve positioning logic

Changed the value added to _alignPositionStreamController from 20 to 15.0 for better alignment behavior in the UI. This adjustment aims to enhance the user experience by refining the positioning of elements on the HomePage.

* Add ValueNotifiers for bottom sheet expansion and collapse states in HomePage

Introduced two ValueNotifiers to track the expansion and collapse states of the bottom sheet. This enhancement allows for better state management and dynamic updates to the UI based on the bottom sheet's current state, improving overall user experience.

* Enhance bottom sheet functionality in HomePage with improved state management

Refactored the bottom sheet and Floating Action Button (FAB) logic by introducing ValueNotifiers to track the expansion and collapse states. Updated the listener for the sheet controller to synchronize the FAB position and bottom sheet state more effectively. This change improves the responsiveness of the UI and enhances user interaction with the bottom sheet.